pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "quyuan"
dynamic = ["version"]
authors = [
    { name = "Tang Ziya", email = "tang@stu.njau.edu.cn" },
    { name = "Applied Bioinformatics Laboratory", email = "abil@ihrc.com" },
]
description = "A simple way to visualize features on human and other species chromosome ideograms, folk of jordanlab/tagore"
readme = "README.md"
keywords = [
    "Human",
    "chromosome",
    "ideogram",
    "visualization",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Environment :: Console",
    "Development Status :: 5 - Production/Stable",
    "Topic :: Scientific/Engineering :: Visualization",
]
requires-python = ">=3.6"
dependencies = [
    "cairosvg",
    "click",
    "colorama",
    "importlib_metadata; python_version<'3.8'",
    "svg.py"
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
quyuan = "quyuan.main:run"
rfmix2tagore = "quyuan.scripts.rfmix2tagore:main"

[project.urls]
homepage = "https://github.com/tcztzy/quyuan"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["quyuan.tests*"]

[tool.setuptools.package-data]
quyuan = ["*.svg.g", "*.json", "*.csv"]

[tool.setuptools_scm]
write_to = "src/quyuan/_version.py"
